Crate rivlib

Read points from Riegl's rxp format.

You need to have RiVLib on your system in order to use this library. Get RiVLib from the Riegl website.

rxp is a data layout specification that Riegl uses for streaming discrete point data. The format is used both to store data in files and to stream data over a network.

Riegl does not publish the specification for the rxp format, so we must use their RiVLib to read rxp data.

Reexports

pub use error::Error;

pub use stream::{Stream, Point};

Modules

error
Our custom error handling.

stream
rxp data streams.

Functions

library_info
Returns information about the library's build version and tag.

library_version
Returns the scanifc library version as a three-tuple.

Type Definitions

Result
Crate-specific results.